Create Graph Database Visualization
This will create a simple web page that visualizes the graph data from the Neo4j database. You should see a visualization of the movies and their relationships.

This code connects to a Neo4j database, runs a Cypher query to fetch movies and their similar movies, and visualizes the results using Neovis.js.